Split IR.hs into Inventory/Selector/Assertion submodules
Re-organize the IR along backend-affinity lines so future backends
(InSpec, etc.) have an obvious place to plug in: Inventory is fully
backend-agnostic, Selector is agnostic in shape, and Assertion carries
kind values whose vocabulary is governed by the backend schema.

PanInfraSpec.IR remains a re-export façade with an unchanged export
list, so every existing import site keeps working without edits. A new
SplitTest verifies the façade re-exports identical types from the
submodules. Closes #37.

Mapping is co-located with Assertion (rather than Selector as the
issue originally proposed) because PlanFile -> Mapping -> Assertion
would otherwise force a Selector <-> Assertion module cycle.
